refactor: Update schedule generator related components directory name and imports
