+++ app/pages/admin/schedules-generator/data/-components/TabCarreras.vue
