+++ app/pages/admin/schedules-generator/data/-components/TabClassrooms.vue
@@ -42,7 +42,7 @@ import { ref } from 'vue'
 import { useSchedulesGeneratorStore } from '~/stores/schedules-generator'
 import { Dialog, DialogContent, DialogDescription, DialogHeader, DialogTitle, DialogTrigger } from '~/components/ui/dialog'
 import Button from '~/components/ui/button/Button.vue'
-import ImportCard from '../../components/ImportCard.vue'
+import ImportCard from '../../-components/ImportCard.vue'
 
 const store = useSchedulesGeneratorStore()
 const isModalOpen = ref(false)
