+++ app/pages/admin/schedules-generator/data/-components/TabCleanups.vue
