+++ app/pages/admin/schedules-generator/data/-components/TabPeriods.vue
