+++ app/pages/admin/schedules-generator/data/-components/TabTeachers.vue
@@ -56,7 +56,7 @@ import Button from '~/components/ui/button/Button.vue'
 import CInputText from '~/components/primitives/form/CInputText.vue'
 import TeacherFormDialog from '~/components/schedules-generator/TeacherFormDialog.vue'
 import DeleteItemDialog from '~/components/dialogs/DeleteItemDialog.vue'
-import ImportCard from '../../components/ImportCard.vue'
+import ImportCard from '../../-components/ImportCard.vue'
 import DataTable from '~/components/partials/datatable/DataTable.vue'
 import { useTableSearch } from '~/composables/Usetablesearch'
 import { fetchTeachers, deleteTeacher } from '~/lib/api/schedules-generator/teachers'
