+++ app/pages/admin/schedules-generator/data/index.vue
@@ -54,16 +54,16 @@ import { useRoute, useRouter } from 'vue-router'
 import { Tabs, TabsList, TabsTrigger } from '~/components/ui/tabs'
 import type { ImportType } from '~/lib/api/schedules-generator/types'
 
-import TabTeachers from './components/TabTeachers.vue'
-import TabCourses from './components/TabCourses.vue'
-import TabCarreras from './components/TabCarreras.vue'
-import TabTeacherCourse from './components/TabTeacherCourse.vue'
-import TabClassrooms from './components/TabClassrooms.vue'
-import TabLabs from './components/TabLabs.vue'
-import TabSections from './components/TabSections.vue'
-import TabSectionLab from './components/TabSectionLab.vue'
-import TabCleanups from './components/TabCleanups.vue'
-import TabPeriods from './components/TabPeriods.vue'
+import TabTeachers from './-components/TabTeachers.vue'
+import TabCourses from './-components/TabCourses.vue'
+import TabCarreras from './-components/TabCarreras.vue'
+import TabTeacherCourse from './-components/TabTeacherCourse.vue'
+import TabClassrooms from './-components/TabClassrooms.vue'
+import TabLabs from './-components/TabLabs.vue'
+import TabSections from './-components/TabSections.vue'
+import TabSectionLab from './-components/TabSectionLab.vue'
+import TabCleanups from './-components/TabCleanups.vue'
+import TabPeriods from './-components/TabPeriods.vue'
 
 const route = useRoute()
 const router = useRouter()
